split TypeGraph utils to own file
