+++ packages/quicktype-core/src/UnionBuilder.ts
@@ -18,7 +18,7 @@ import {
     isPrimitiveTypeKind
 } from "./Type";
 import { type TypeBuilder } from "./Type/TypeBuilder";
-import { type TypeRef } from "./Type/TypeGraph";
+import { type TypeRef } from "./Type/TypeGraphUtils";
 import { matchTypeExhaustive } from "./Type/TypeUtils";
 
 // FIXME: This interface is badly designed.  All the properties
